+++ src/js/core/form-custom.js
@@ -7,6 +7,8 @@ export default function (UIkit) {
 
         mixins: [Class],
 
+        args: 'target',
+
         props: {
             target: Boolean
         },